pyproject.toml
[project]
name = "sidecar-eq"
version = "1.1.1"
description = "Educational audio player with thermometer-style EQ interface for music analysis and per-track customization."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.7",
  "librosa>=0.10",
  "soundfile>=0.12",
  "numpy",
  "scipy",
  "pyqtgraph",
  "python-dotenv>=1.0",
  "plexapi>=4.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sidecar_eq"]

[project.optional-dependencies]
yt = ["yt-dlp"]
dev = [
  "black>=24.0",
  "ruff>=0.1.0",
  "mypy>=1.0",
  "pytest>=7.0",
  "pytest-qt>=4.0",
  "isort>=5.12",
]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["sidecar_eq"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # line too long (handled by black)
  "B008",  # do not perform function calls in argument defaults
  "C901",  # too complex
]
exclude = [
  ".git",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "__pycache__",
  "build",
  "dist",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually
ignore_missing_imports = true
